requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pydantic==2.11.7
python-dotenv==1.1.1
psycopg[binary]==3.2.10
alembic==1.16.5
redis==5.2.1
